requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
asyncio==3.4.3
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
discord==2.3.2
discord.py==2.4.0
Flask==3.0.3
frozenlist==1.5.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
multidict==6.1.0
propcache==0.2.0
py-cord==2.6.1
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3
Werkzeug==3.1.1
yarl==1.17.1
